Correction on Migrations

+++ src/main/resources/db/migration/V1.0__init.sql
@@ -1,14 +1,14 @@
 create table speciality (
-   id                bigserial not null,
+   id                serial not null,
    description       varchar(255),
 
    constraint pk_speciality primary key (id)
 );
 
 create table doctor (
-   id               bigserial not null,
+   id               serial not null,
    name             varchar(255),
-   speciality       bigserial not null,
+   speciality       int4 not null,
 
    constraint pk_doctor primary key (id)
 );
@@ -16,7 +16,7 @@ create table doctor (
 alter table if exists doctor add constraint FK_DOCTOR_ON_SPECIALITY foreign key (speciality) references speciality (id);
 
 create table patient (
-   id               bigserial not null,
+   id               serial not null,
    name             varchar(255) not null,
    age              integer,
 
@@ -24,10 +24,10 @@ create table patient (
 );
 
 create table consult (
-   id               bigserial not null,
-   patient          bigserial not null,
-   doctor           bigserial not null,
-   speciality       bigserial not null,
+   id               serial not null,
+   patient          int4 not null,
+   doctor           int4 not null,
+   speciality       int4 not null,
    
    constraint pk_consult primary key (id)
 );
@@ -39,22 +39,22 @@ alter table if exists consult add constraint FK_CONSULT_ON_PATIENT foreign key (
 alter table if exists consult add constraint FK_CONSULT_ON_SPECIALITY foreign key (speciality) references speciality (id);
 
 create table pathology (
-   id               bigserial not null,
+   id               serial not null,
    description      varchar(255),
 
    constraint pk_pathology primary key (id)
 );
 
 create table symptom (
-    id               bigserial not null,
+    id               serial not null,
     description      varchar(255),
 
     constraint pk_symptom primary key (id)
 );
 
 create table symptom_pathology(
- pathology_id         bigserial,
- symptom_id           bigserial,
+ pathology_id         int4,
+ symptom_id           int4,
 
  constraint pk_symptom_pathology primary key (pathology_id, symptom_id)
 );
@@ -64,8 +64,8 @@ alter table if exists symptom_pathology add constraint FK_PATHOLOGY_ON_SYMPTOM_P
 
 
 create table patient_pathology(
- pathology_id          bigserial,
- patient_id            bigserial,
+ pathology_id          int4,
+ patient_id            int4,
 
  constraint pk_patient_pathology primary key (pathology_id, patient_id)
 );
